+++ Changes.md
@@ -38,6 +38,9 @@ Unreleased
   types](https://v2.ocaml.org/releases/5.0/htmlman/types.html#sss:typexpr-aliased-recursive)
 - [melange.ppx]: Deprecate `bs.*` attributes in favor of `mel.*`
   ([#662](https://github.com/melange-re/melange/pull/662))
+- [melange]: Fix field access code generation when `open`in inline functor
+  applications ([#661](https://github.com/melange-re/melange/pull/661),
+  [#664](https://github.com/melange-re/melange/pull/664))
 
 1.0.0 2023-05-31
 ---------------